Move event loop to loop.{cc,h}
